General cleanup per brianwilkerson; dead code, utility methods
New occurrences file and tests, disabled. Delete lots of commented out
dead code related to navigation before refactoring.
