+++ angular_analyzer_plugin/lib/src/navigation.dart
@@ -1,7 +1,3 @@
-//import 'package:analysis_server/plugin/analysis/navigation/navigation_core.dart';
-//import 'package:analysis_server/plugin/analysis/occurrences/occurrences_core.dart';
-//import 'package:analysis_server/protocol/protocol_generated.dart' as protocol;
-//import 'package:analysis_server/plugin/protocol/protocol_dart.dart' as protocol;
 import 'package:analyzer_plugin/protocol/protocol_common.dart' as protocol;
 import 'package:analyzer_plugin/protocol/protocol_constants.dart' as protocol;
 import 'package:analyzer_plugin/utilities/navigation/navigation.dart';
@@ -54,7 +50,7 @@ class AngularNavigation implements NavigationContributor {
   void _addDirectiveRegions(NavigationCollector collector,
       AbstractDirective directive, SourceRange targetRange) {
     for (final input in directive.inputs) {
-      if (!rangesOverlap(input.setterRange, targetRange)) {
+      if (!isTargeted(input.setterRange, targetRange: targetRange)) {
         continue;
       }
       final setter = input.setter;
@@ -84,7 +80,7 @@ class AngularNavigation implements NavigationContributor {
   void _addTemplateRegions(NavigationCollector collector, Template template,
       SourceRange targetRange) {
     for (final resolvedRange in template.ranges) {
-      if (!rangesOverlap(resolvedRange.range, targetRange)) {
+      if (!isTargeted(resolvedRange.range, targetRange: targetRange)) {
         continue;
       }
 
@@ -115,7 +111,7 @@ class AngularNavigation implements NavigationContributor {
   void _addViewRegions(
       NavigationCollector collector, View view, SourceRange targetRange) {
     if (view.templateUriSource != null &&
-        rangesOverlap(view.templateUrlRange, targetRange)) {
+        isTargeted(view.templateUrlRange, targetRange: targetRange)) {
       collector.addRegion(
           view.templateUrlRange.offset,
           view.templateUrlRange.length,
@@ -124,108 +120,8 @@ class AngularNavigation implements NavigationContributor {
     }
   }
 
-  bool rangesOverlap(SourceRange a, SourceRange b) =>
+  /// A null target range indicates everything is targeted. Otherwise, intersect
+  bool isTargeted(SourceRange toTest, {SourceRange targetRange}) =>
       // <a><b></b></a> or <a><b></a></b>, but not <a></a><b></b>.
-      a == null || b == null || a.contains(b.offset) || b.contains(a.offset);
+      targetRange == null || targetRange.intersects(toTest);
 }
-
-//class AngularOccurrencesContributor implements OccurrencesContributor {
-//  @override
-//  void computeOccurrences(
-//      OccurrencesCollector collector, AnalysisContext context, Source source) {
-//    //List<Source> librarySources = context.getLibrariesContaining(source);
-//    //for (Source librarySource in librarySources) {
-//    //  // directives
-//    //  {
-//    //    List<AbstractDirective> directives = context.getResult(
-//    //        new LibrarySpecificUnit(librarySource, source), DIRECTIVES_IN_UNIT);
-//    //    for (AbstractDirective directive in directives) {
-//    //      _addDirectiveOccurrences(collector, directive);
-//    //    }
-//    //  }
-//    //  // templates
-//    //  {
-//    //    List<Template> templates = context.getResult(
-//    //        new LibrarySpecificUnit(librarySource, source), DART_TEMPLATES);
-//    //    for (Template template in templates) {
-//    //      _addTemplateOccurrences(collector, template);
-//    //    }
-//    //  }
-//    //}
-//  }
-//
-//  void addDirectiveOccurrences(
-//      OccurrencesCollector collector, AbstractDirective directive) {
-//    final elementsOffsets = <engine.PropertyAccessorElement, List<int>>{};
-//    for (final input in directive.inputs) {
-//      final setter = input.setter;
-//      if (setter == null) {
-//        continue;
-//      }
-//      var offsets = elementsOffsets[setter];
-//      if (offsets == null) {
-//        offsets = <int>[setter.nameOffset];
-//        elementsOffsets[setter] = offsets;
-//      }
-//      offsets.add(input.setterRange.offset);
-//    }
-//    // convert map into Occurrences
-//    elementsOffsets.forEach((setter, offsets) {
-//      final protocolElement = _newProtocolElementForEngine(setter);
-//      final length = protocolElement.location.length;
-//      final occurrences =
-//          new protocol.Occurrences(protocolElement, offsets, length);
-//      collector.addOccurrences(occurrences);
-//    });
-//  }
-//
-//  void addTemplateOccurrences(
-//      OccurrencesCollector collector, Template template) {
-//    final elementsOffsets = <AngularElement, List<int>>{};
-//    for (final resolvedRange in template.ranges) {
-//      final element = resolvedRange.element;
-//      var offsets = elementsOffsets[element];
-//      if (offsets == null) {
-//        offsets = <int>[element.nameOffset];
-//        elementsOffsets[element] = offsets;
-//      }
-//      offsets.add(resolvedRange.range.offset);
-//    }
-//    // convert map into Occurrences
-//    elementsOffsets.forEach((angularElement, offsets) {
-//      final length = angularElement.nameLength;
-//      final protocolElement = _newProtocolElement(angularElement);
-//      final occurrences =
-//          new protocol.Occurrences(protocolElement, offsets, length);
-//      collector.addOccurrences(occurrences);
-//    });
-//  }
-//
-//  engine.Element _canonicalizeElement(engine.Element element) {
-//    var canonical = element;
-//    if (canonical is engine.PropertyAccessorElement) {
-//      canonical = (canonical as engine.PropertyAccessorElement).variable;
-//    }
-//    if (canonical is Member) {
-//      canonical = (canonical as Member).baseElement;
-//    }
-//    return canonical;
-//  }
-//
-//  protocol.Element _newProtocolElement(AngularElement angularElement) {
-//    final name = angularElement.name;
-//    final length = name.length;
-//    if (angularElement is DartElement) {
-//      final dartElement = angularElement.element;
-//      return _newProtocolElementForEngine(dartElement);
-//    }
-//    return new protocol.Element(protocol.ElementKind.UNKNOWN, name, 0,
-//        location: new protocol.Location(angularElement.source.fullName,
-//            angularElement.nameOffset, length, -1, -1));
-//  }
-//
-//  protocol.Element _newProtocolElementForEngine(engine.Element dartElement) {
-//    final cannonical = _canonicalizeElement(dartElement);
-//    return protocol.convertElement(cannonical);
-//  }
-//}
