+++ angular_analyzer_plugin/test/navigation_test.dart
@@ -1,7 +1,6 @@
 import 'dart:async';
 
 import 'package:analyzer_plugin/utilities/navigation/navigation.dart';
-//import 'package:analysis_server/plugin/analysis/occurrences/occurrences_core.dart';
 import 'package:analyzer_plugin/protocol/protocol_common.dart' as protocol;
 import 'package:analyzer/error/error.dart';
 import 'package:analyzer/error/listener.dart';
@@ -18,19 +17,8 @@ import 'abstract_angular.dart';
 
 void main() {
   defineReflectiveSuite(() {
-    // TODO get these working again in the latest SDK
     defineReflectiveTests(AngularNavigationTest);
-    //defineReflectiveTests(AngularOccurrencesContributorTest);
-    defineReflectiveTests(EmptyTest);
   });
-}
-
-@reflectiveTest
-class EmptyTest {
-  // ignore: non_constant_identifier_names
-  void test_soTheSuitePasses() {
-    expect(null, isNull);
-  }
 }
 
 @reflectiveTest
@@ -384,95 +372,6 @@ class TestComponent {
     _findRegion(offset, str.length);
   }
 }
-
-//@reflectiveTest
-//class AngularOccurrencesContributorTest extends AbstractAngularTest {
-//  String code;
-//
-//  OccurrencesCollector collector = new OccurrencesCollectorMock();
-//  List<protocol.Occurrences> occurrencesList = <protocol.Occurrences>[];
-//
-//  protocol.Occurrences occurrences;
-//
-//  @override
-//  void setUp() {
-//    super.setUp();
-//    when(collector.addOccurrences(anyObject)).thenInvoke(occurrencesList.add);
-//  }
-//
-//  // ignore: non_constant_identifier_names
-//  void test_dart_templates() {
-//    code = r'''
-//import 'package:angular/src/core/metadata.dart';
-//
-//@Component(selector: 'text-panel', inputs: const ['text: my-text'])
-//@View(template: r"<div>some text</div>")
-//class TextPanel {
-//  String text; // 1
-//}
-//
-//@Component(selector: 'UserPanel')
-//@View(template: r"""
-//<div>
-//  <text-panel [my-text]='user.value'></text-panel> // cl
-//</div>
-//""", directives: [TextPanel])
-//class UserPanel {
-//  ObjectContainer<String> user; // 2
-//}
-//
-//class ObjectContainer<T> {
-//  T value; // 3
-//}
-//''';
-//    final source = newSource('/test.dart', code);
-//    //LibrarySpecificUnit target = new LibrarySpecificUnit(source, source);
-//    //computeResult(target, DART_TEMPLATES);
-//    // compute navigation regions
-//    new AngularOccurrencesContributor()
-//        .computeOccurrences(collector, null, source);
-//    // "text" field
-//    {
-//      _findOccurrences(code.indexOf('text: my-text'));
-//      expect(occurrences.element.name, 'text');
-//      expect(occurrences.length, 'text'.length);
-//      expect(occurrences.offsets, contains(code.indexOf('text; // 1')));
-//    }
-//    // "text-panel" component
-//    {
-//      _findOccurrences(code.indexOf("text-panel', "));
-//      expect(occurrences.element.name, 'text-panel');
-//      expect(occurrences.length, 'text-panel'.length);
-//      expect(occurrences.offsets, contains(code.indexOf("text-panel [")));
-//      expect(occurrences.offsets, contains(code.indexOf("text-panel> // cl")));
-//    }
-//    // "user" field
-//    {
-//      _findOccurrences(code.indexOf("user.value'><"));
-//      expect(occurrences.element.name, 'user');
-//      expect(occurrences.length, 'user'.length);
-//      expect(occurrences.offsets, contains(code.indexOf('user; // 2')));
-//    }
-//    // "value" field
-//    {
-//      _findOccurrences(code.indexOf("value'><"));
-//      expect(occurrences.element.name, 'value');
-//      expect(occurrences.length, 'value'.length);
-//      expect(occurrences.offsets, contains(code.indexOf('value; // 3')));
-//    }
-//  }
-//
-//  void _findOccurrences(int offset) {
-//    for (final occurrences in occurrencesList) {
-//      if (occurrences.offsets.contains(offset)) {
-//        this.occurrences = occurrences;
-//        return;
-//      }
-//    }
-//    final listStr = occurrencesList.join('\n');
-//    fail('Unable to find occurrences at $offset in $listStr');
-//  }
-//}
 
 /// Instances of the class [GatheringErrorListener] implement an error listener
 /// that collects all of the errors passed to it for later examination.
