+++ angular_analyzer_plugin/test/test_all.dart
@@ -3,6 +3,7 @@ import 'package:test_reflective_loader/test_reflective_loader.dart';
 import 'angular_driver_test.dart' as angular_driver_test;
 import 'completion_contributor_test.dart' as completion_contributor_test;
 import 'navigation_test.dart' as navigation_test;
+import 'occurrences_test.dart' as occurrences_test;
 import 'offsetting_constant_value_visitor_test.dart'
     as offsetting_constant_value_visitor_test;
 import 'plugin_test.dart' as plugin_test;
@@ -19,6 +20,7 @@ void main() {
     angular_driver_test.main();
     offsetting_constant_value_visitor_test.main();
     navigation_test.main();
+    occurrences_test.main();
     completion_contributor_test.main();
   }, name: 'Angular Plugin tests');
 }
